:recycle: : 메소드 분리 (#104)
