+++ Match-Api/src/main/java/com/example/matchapi/order/helper/OrderHelper.java
@@ -1,29 +1,21 @@
 package com.example.matchapi.order.helper;
 
-import com.example.matchapi.portone.dto.PaymentReq;
+import com.example.matchapi.order.dto.OrderRes;
 import com.example.matchcommon.annotation.Helper;
-import com.example.matchcommon.exception.BaseException;
-import com.example.matchcommon.properties.NicePayProperties;
 import com.example.matchdomain.donation.adaptor.DonationAdaptor;
 import com.example.matchdomain.donation.entity.DonationUser;
 import com.example.matchdomain.donation.entity.enums.PayMethod;
 import com.example.matchdomain.donation.entity.flameEnum.Adjective;
 import com.example.matchdomain.donation.entity.flameEnum.AdjectiveFlame;
-import com.example.matchdomain.donation.repository.DonationUserRepository;
-import com.example.matchdomain.project.entity.Project;
 import com.example.matchdomain.user.entity.User;
-import com.example.matchinfrastructure.pay.nice.client.NiceAuthFeignClient;
-import com.example.matchinfrastructure.pay.nice.dto.NicePayCancelRequest;
-import com.siot.IamportRestClient.response.Payment;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.springframework.http.HttpStatus;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Base64;
 import java.util.List;
 import java.util.Random;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 import static com.example.matchcommon.constants.MatchStatic.BASIC;
@@ -31,8 +23,6 @@ import static com.example.matchcommon.constants.MatchStatic.BASIC;
 @Helper
 @RequiredArgsConstructor
 public class OrderHelper {
-    private final NicePayProperties nicePayProperties;
-    private final NiceAuthFeignClient niceAuthFeignClient;
     private final DonationAdaptor donationAdaptor;
 
     public PayMethod getPayMethod(String value) {
@@ -42,10 +32,6 @@ public class OrderHelper {
             }
         }
         return null;
-    }
-
-    public String getNicePaymentAuthorizationHeader() {
-        return BASIC + Base64.getEncoder().encodeToString((nicePayProperties.getClient() + ":" + nicePayProperties.getSecret()).getBytes());
     }
 
     public List<String> getInherenceName(List<DonationUser> donationUsers){
@@ -88,4 +74,29 @@ public class OrderHelper {
         return type + LocalDateTime.now().format(DateTimeFormatter.ofPattern("yy.MM.dd.HH:mm")) + "-" + randomStr;
     }
 
+    public OrderRes.CreateInherenceDto createInherence(User user) {
+
+        String flameName = createFlameName(user);
+
+        String inherenceNumber = createRandomUUID();
+
+        return new OrderRes.CreateInherenceDto(flameName, inherenceNumber);
+    }
+
+    public String createRandomUUID() {
+        return LocalDateTime.now().format(DateTimeFormatter.ofPattern("yy.MM.dd.HH:mm")) + "." + UUID.randomUUID();
+    }
+
+    public String formatString(String input, int length) {
+        StringBuilder formatted = new StringBuilder();
+
+        for (int i = 0; i < input.length(); i++) {
+            if (i > 0 && i % length == 0) {
+                formatted.append('-');
+            }
+            formatted.append(input.charAt(i));
+        }
+
+        return formatted.toString();
+    }
 }
